+++ src/prep.py
@@ -2,25 +2,20 @@
 from src.args import Args
 from src.console import console
 from src.exceptions import *  # noqa: F403
-from src.trackers.PTP import PTP  # noqa F401
-from src.trackers.BLU import BLU  # noqa F401
-from src.trackers.AITHER import AITHER  # noqa F401
-from src.trackers.LST import LST  # noqa F401
-from src.trackers.OE import OE  # noqa F401
-from src.trackers.HDB import HDB  # noqa F401
-from src.trackers.TIK import TIK  # noqa F401
+from src.trackers.PTP import PTP
 from src.trackers.COMMON import COMMON
 from src.clients import Clients
 from data.config import config
 from src.uphelper import UploadHelper
 from src.trackersetup import TRACKER_SETUP, tracker_class_map
+from src.takescreens import disc_screenshots, dvd_screenshots, screenshots
+from src.tvmaze import search_tvmaze
+from src.imdb import get_imdb_aka_api, get_imdb_info_api, search_imdb, imdb_other_meta
+from src.trackermeta import update_metadata_from_tracker
 
 try:
     import traceback
     from src.discparse import DiscParse
-    import multiprocessing
-    from multiprocessing import get_context
-    from tqdm import tqdm
     import os
     import re
     import math
@@ -31,32 +26,21 @@ try:
     from pathlib import Path
     import urllib
     import urllib.parse
-    import ffmpeg
     import random
     import json
     import glob
     import requests
-    import pyimgbox
     from pymediainfo import MediaInfo
     import tmdbsimple as tmdb
     from datetime import datetime
     from difflib import SequenceMatcher
     import torf
     from torf import Torrent
-    import base64
     import time
     import anitopy
-    import shutil
-    from imdb import Cinemagoer
     import itertools
     import cli_ui
-    from rich.progress import Progress, TextColumn, BarColumn, TimeRemainingColumn  # noqa F401
-    import platform
     import aiohttp
-    from PIL import Image
-    import io
-    from io import BytesIO
-    import sys
 except ModuleNotFoundError:
     console.print(traceback.print_exc())
     console.print('[bold red]Missing Module Found. Please reinstall required dependancies.')
@@ -79,346 +63,6 @@ class Prep():
         self.config = config
         self.img_host = img_host.lower()
         tmdb.API_KEY = config['DEFAULT']['tmdb_api']
-
-    async def prompt_user_for_confirmation(self, message: str) -> bool:
-        try:
-            response = input(f"{message} (Y/n): ").strip().lower()
-            if response in ["y", "yes", ""]:
-                return True
-            return False
-        except EOFError:
-            sys.exit(1)
-
-    async def check_images_concurrently(self, imagelist, meta):
-        approved_image_hosts = ['ptpimg', 'imgbox', 'imgbb']
-        invalid_host_found = False  # Track if any image is on a non-approved host
-
-        # Ensure meta['image_sizes'] exists
-        if 'image_sizes' not in meta:
-            meta['image_sizes'] = {}
-
-        # Map fixed resolution names to vertical resolutions
-        resolution_map = {
-            '8640p': 8640,
-            '4320p': 4320,
-            '2160p': 2160,
-            '1440p': 1440,
-            '1080p': 1080,
-            '1080i': 1080,
-            '720p': 720,
-            '576p': 576,
-            '576i': 576,
-            '480p': 480,
-            '480i': 480,
-        }
-
-        # Get expected vertical resolution
-        expected_resolution_name = meta.get('resolution', None)
-        expected_vertical_resolution = resolution_map.get(expected_resolution_name, None)
-
-        # If no valid resolution is found, skip processing
-        if expected_vertical_resolution is None:
-            console.print("[red]Meta resolution is invalid or missing. Skipping all images.[/red]")
-            return []
-
-        # Function to check each image's URL, host, and log resolution
-        async def check_and_collect(image_dict):
-            img_url = image_dict.get('raw_url')
-            if not img_url:
-                return None
-
-            if "ptpimg.me" in img_url and img_url.startswith("http://"):
-                img_url = img_url.replace("http://", "https://")
-                image_dict['raw_url'] = img_url
-                image_dict['web_url'] = img_url
-
-            # Verify the image link
-            if await self.check_image_link(img_url):
-                # Check if the image is hosted on an approved image host
-                if not any(host in img_url for host in approved_image_hosts):
-                    nonlocal invalid_host_found
-                    invalid_host_found = True  # Mark that we found an invalid host
-
-                async with aiohttp.ClientSession() as session:
-                    async with session.get(img_url) as response:
-                        if response.status == 200:
-                            image_content = await response.read()
-
-                            try:
-                                image = Image.open(BytesIO(image_content))
-                                vertical_resolution = image.height
-                                lower_bound = expected_vertical_resolution * 0.70  # 30% below
-                                if meta['is_disc'] == "DVD":
-                                    upper_bound = expected_vertical_resolution * 1.30
-                                else:
-                                    upper_bound = expected_vertical_resolution * 1.00
-
-                                if not (lower_bound <= vertical_resolution <= upper_bound):
-                                    console.print(
-                                        f"[red]Image {img_url} resolution ({vertical_resolution}p) "
-                                        f"is outside the allowed range ({int(lower_bound)}-{int(upper_bound)}p). Skipping.[/red]"
-                                    )
-                                    return None
-
-                                meta['image_sizes'][img_url] = len(image_content)
-                                console.print(
-                                    f"Valid image {img_url} with resolution {image.width}x{image.height} "
-                                    f"and size {len(image_content) / 1024:.2f} KiB"
-                                )
-                            except Exception as e:
-                                console.print(f"[red]Failed to process image {img_url}: {e}")
-                                return None
-                        else:
-                            console.print(f"[red]Failed to fetch image {img_url}. Skipping.")
-
-                return image_dict
-            else:
-                return None
-
-        # Run image verification concurrently
-        tasks = [check_and_collect(image_dict) for image_dict in imagelist]
-        results = await asyncio.gather(*tasks)
-
-        # Collect valid images
-        valid_images = [image for image in results if image is not None]
-
-        # Convert default_trackers string into a list
-        default_trackers = self.config['TRACKERS'].get('default_trackers', '')
-        trackers_list = [tracker.strip() for tracker in default_trackers.split(',')]
-
-        # Ensure meta['trackers'] is a list
-        if meta.get('trackers') is not None:
-            if isinstance(meta.get('trackers', ''), str):
-                meta['trackers'] = [tracker.strip() for tracker in meta['trackers'].split(',')]
-            if 'MTV' in meta.get('trackers', []):
-                if invalid_host_found:
-                    console.print(
-                        "[red]Warning: Some images are not hosted on an MTV-approved image host. MTV will need new images later.[/red]"
-                    )
-        # Issue warning if any valid image is on an unapproved host and MTV is in the trackers list
-        elif 'MTV' in trackers_list:
-            if invalid_host_found:
-                console.print("[red]Warning: Some images are not hosted on an MTV-approved image host. MTV will need new images later.[/red]")
-
-        return valid_images
-
-    async def check_image_link(self, url):
-        async with aiohttp.ClientSession() as session:
-            try:
-                async with session.get(url) as response:
-                    if response.status == 200:
-                        content_type = response.headers.get('Content-Type', '').lower()
-                        if 'image' in content_type:
-                            # Attempt to load the image
-                            image_data = await response.read()
-                            try:
-                                image = Image.open(io.BytesIO(image_data))
-                                image.verify()  # This will check if the image is broken
-                                console.print(f"[green]Image verified successfully: {url}[/green]")
-                                return True
-                            except (IOError, SyntaxError) as e:  # noqa #F841
-                                console.print(f"[red]Image verification failed (corrupt image): {url}[/red]")
-                                return False
-                        else:
-                            console.print(f"[red]Content type is not an image: {url}[/red]")
-                            return False
-                    else:
-                        console.print(f"[red]Failed to retrieve image: {url} (status code: {response.status})[/red]")
-                        return False
-            except Exception as e:
-                console.print(f"[red]Exception occurred while checking image: {url} - {str(e)}[/red]")
-                return False
-
-    async def update_meta_with_unit3d_data(self, meta, tracker_data, tracker_name):
-        # Unpack the expected 9 elements, ignoring any additional ones
-        tmdb, imdb, tvdb, mal, desc, category, infohash, imagelist, filename, *rest = tracker_data
-
-        if tmdb not in [None, '0']:
-            meta['tmdb_manual'] = tmdb
-        if imdb not in [None, '0']:
-            meta['imdb'] = str(imdb).zfill(7)
-        if tvdb not in [None, '0']:
-            meta['tvdb_id'] = tvdb
-        if mal not in [None, '0']:
-            meta['mal'] = mal
-        if desc not in [None, '0', '']:
-            meta['description'] = desc
-            with open(f"{meta['base_dir']}/tmp/{meta['uuid']}/DESCRIPTION.txt", 'w', newline="", encoding='utf8') as description:
-                description.write((desc or "") + "\n")
-        if category.upper() in ['MOVIE', 'TV SHOW', 'FANRES']:
-            meta['category'] = 'TV' if category.upper() == 'TV SHOW' else category.upper()
-
-        if not meta.get('image_list'):  # Only handle images if image_list is not already populated
-            if imagelist:  # Ensure imagelist is not empty before setting
-                valid_images = await self.check_images_concurrently(imagelist, meta)
-                if valid_images:
-                    meta['image_list'] = valid_images
-                    if meta.get('image_list'):  # Double-check if image_list is set before handling it
-                        if not (meta.get('blu') or meta.get('aither') or meta.get('lst') or meta.get('oe') or meta.get('tik')) or meta['unattended']:
-                            await self.handle_image_list(meta, tracker_name)
-
-        if filename:
-            meta[f'{tracker_name.lower()}_filename'] = filename
-
-        console.print(f"[green]{tracker_name} data successfully updated in meta[/green]")
-
-    async def update_metadata_from_tracker(self, tracker_name, tracker_instance, meta, search_term, search_file_folder):
-        tracker_key = tracker_name.lower()
-        manual_key = f"{tracker_key}_manual"
-        found_match = False
-
-        if tracker_name in ["BLU", "AITHER", "LST", "OE", "TIK"]:
-            if meta.get(tracker_key) is not None:
-                console.print(f"[cyan]{tracker_name} ID found in meta, reusing existing ID: {meta[tracker_key]}[/cyan]")
-                tracker_data = await COMMON(self.config).unit3d_torrent_info(
-                    tracker_name,
-                    tracker_instance.torrent_url,
-                    tracker_instance.search_url,
-                    meta,
-                    id=meta[tracker_key]
-                )
-            else:
-                console.print(f"[yellow]No ID found in meta for {tracker_name}, searching by file name[/yellow]")
-                tracker_data = await COMMON(self.config).unit3d_torrent_info(
-                    tracker_name,
-                    tracker_instance.torrent_url,
-                    tracker_instance.search_url,
-                    meta,
-                    file_name=search_term
-                )
-
-            if any(item not in [None, '0'] for item in tracker_data[:3]):  # Check for valid tmdb, imdb, or tvdb
-                console.print(f"[green]Valid data found on {tracker_name}, setting meta values[/green]")
-                await self.update_meta_with_unit3d_data(meta, tracker_data, tracker_name)
-                found_match = True
-            else:
-                console.print(f"[yellow]No valid data found on {tracker_name}[/yellow]")
-                found_match = False
-
-        elif tracker_name == "PTP":
-            imdb_id = None
-            if meta.get('ptp') is None:
-                imdb_id, ptp_torrent_id, ptp_torrent_hash = await tracker_instance.get_ptp_id_imdb(search_term, search_file_folder, meta)
-                if ptp_torrent_id:
-                    meta['imdb'] = str(imdb_id).zfill(7) if imdb_id else None
-                    console.print(f"[green]{tracker_name} IMDb ID found: tt{meta['imdb']}[/green]")
-
-                    if not meta['unattended']:
-                        if await self.prompt_user_for_confirmation("Do you want to use this ID data from PTP?"):
-                            found_match = True
-                            meta['ptp'] = ptp_torrent_id
-                            ptp_desc, ptp_imagelist = await tracker_instance.get_ptp_description(ptp_torrent_id, meta, meta.get('is_disc', False))
-                            meta['description'] = ptp_desc
-                            with open(f"{meta['base_dir']}/tmp/{meta['uuid']}/DESCRIPTION.txt", 'w', newline="", encoding='utf8') as description:
-                                description.write((ptp_desc or "") + "\n")
-
-                            if not meta.get('image_list'):
-                                valid_images = await self.check_images_concurrently(ptp_imagelist, meta)
-                                if valid_images:
-                                    meta['image_list'] = valid_images
-                                    await self.handle_image_list(meta, tracker_name)
-
-                        else:
-                            found_match = False
-                            meta['imdb'] = None
-
-                    else:
-                        found_match = True
-                        ptp_desc, ptp_imagelist = await tracker_instance.get_ptp_description(ptp_torrent_id, meta, meta.get('is_disc', False))
-                        meta['description'] = ptp_desc
-                        with open(f"{meta['base_dir']}/tmp/{meta['uuid']}/DESCRIPTION.txt", 'w', newline="", encoding='utf8') as description:
-                            description.write((ptp_desc or "") + "\n")
-                        meta['saved_description'] = True
-
-                        if not meta.get('image_list'):
-                            valid_images = await self.check_images_concurrently(ptp_imagelist, meta)
-                            if valid_images:
-                                meta['image_list'] = valid_images
-                else:
-                    console.print("[yellow]Skipping PTP as no match found[/yellow]")
-                    found_match = False
-
-            else:
-                ptp_torrent_id = meta['ptp']
-                console.print("[cyan]Using specified PTP ID to get IMDb ID[/cyan]")
-                imdb_id, _, meta['ext_torrenthash'] = await tracker_instance.get_imdb_from_torrent_id(ptp_torrent_id)
-                if imdb_id:
-                    meta['imdb'] = str(imdb_id).zfill(7)
-                    console.print(f"[green]IMDb ID found: tt{meta['imdb']}[/green]")
-                    found_match = True
-                    meta['skipit'] = True
-                    ptp_desc, ptp_imagelist = await tracker_instance.get_ptp_description(meta['ptp'], meta, meta.get('is_disc', False))
-                    meta['description'] = ptp_desc
-                    with open(f"{meta['base_dir']}/tmp/{meta['uuid']}/DESCRIPTION.txt", 'w', newline="", encoding='utf8') as description:
-                        description.write(ptp_desc + "\n")
-                    meta['saved_description'] = True
-                    if not meta.get('image_list'):  # Only handle images if image_list is not already populated
-                        valid_images = await self.check_images_concurrently(ptp_imagelist, meta)
-                        if valid_images:
-                            meta['image_list'] = valid_images
-                            console.print("[green]PTP images added to metadata.[/green]")
-                else:
-                    console.print(f"[yellow]Could not find IMDb ID using PTP ID: {ptp_torrent_id}[/yellow]")
-                    found_match = False
-
-        elif tracker_name == "HDB":
-            if meta.get('hdb') is not None:
-                meta[manual_key] = meta[tracker_key]
-                console.print(f"[cyan]{tracker_name} ID found in meta, reusing existing ID: {meta[tracker_key]}[/cyan]")
-
-                # Use get_info_from_torrent_id function if ID is found in meta
-                imdb, tvdb_id, hdb_name, meta['ext_torrenthash'] = await tracker_instance.get_info_from_torrent_id(meta[tracker_key])
-
-                meta['tvdb_id'] = str(tvdb_id) if tvdb_id else meta.get('tvdb_id')
-                meta['hdb_name'] = hdb_name
-                found_match = True
-
-                # Skip user confirmation if searching by ID
-                console.print(f"[green]{tracker_name} data found: IMDb ID: {imdb}, TVDb ID: {meta['tvdb_id']}, HDB Name: {meta['hdb_name']}[/green]")
-            else:
-                console.print("[yellow]No ID found in meta for HDB, searching by file name[/yellow]")
-
-                # Use search_filename function if ID is not found in meta
-                imdb, tvdb_id, hdb_name, meta['ext_torrenthash'], tracker_id = await tracker_instance.search_filename(search_term, search_file_folder, meta)
-
-                meta['tvdb_id'] = str(tvdb_id) if tvdb_id else meta.get('tvdb_id')
-                meta['hdb_name'] = hdb_name
-                if tracker_id:
-                    meta[tracker_key] = tracker_id
-                found_match = True
-
-                if found_match:
-                    if imdb or tvdb_id or hdb_name:
-                        console.print(f"[green]{tracker_name} data found: IMDb ID: {imdb}, TVDb ID: {meta['tvdb_id']}, HDB Name: {meta['hdb_name']}[/green]")
-                        if await self.prompt_user_for_confirmation(f"Do you want to use the ID's found on {tracker_name}?"):
-                            console.print(f"[green]{tracker_name} data retained.[/green]")
-                        else:
-                            console.print(f"[yellow]{tracker_name} data discarded.[/yellow]")
-                            meta[tracker_key] = None
-                            meta['tvdb_id'] = None
-                            meta['hdb_name'] = None
-                            found_match = False
-                    else:
-                        found_match = False
-
-        return meta, found_match
-
-    async def handle_image_list(self, meta, tracker_name):
-        if meta.get('image_list'):
-            console.print(f"[cyan]Found the following images from {tracker_name}:")
-            for img in meta['image_list']:
-                console.print(f"[blue]{img}[/blue]")
-
-            if meta['unattended']:
-                keep_images = True
-            else:
-                keep_images = await self.prompt_user_for_confirmation(f"Do you want to keep the images found on {tracker_name}?")
-                if not keep_images:
-                    meta['image_list'] = []
-                    meta['image_sizes'] = {}
-                    console.print(f"[yellow]Images discarded from {tracker_name}.")
-                else:
-                    console.print(f"[green]Images retained from {tracker_name}.")
 
     async def gather_prep(self, meta, mode):
         meta['cutoff'] = int(self.config['DEFAULT'].get('cutoff_screens', 3))
@@ -590,7 +234,7 @@ class Prep():
 
                     tracker_instance = tracker_class(config=self.config)
                     try:
-                        updated_meta, match = await self.update_metadata_from_tracker(
+                        updated_meta, match = await update_metadata_from_tracker(
                             tracker_name, tracker_instance, meta, search_term, search_file_folder
                         )
                         if match:
@@ -646,19 +290,19 @@ class Prep():
 
         # If no tmdb, use imdb for meta
         if int(meta['tmdb']) == 0:
-            meta = await self.imdb_other_meta(meta)
+            meta = await imdb_other_meta(meta)
         else:
             meta = await self.tmdb_other_meta(meta)
         # Search tvmaze
         if meta['category'] == "TV":
-            meta['tvmaze_id'], meta['imdb_id'], meta['tvdb_id'] = await self.search_tvmaze(filename, meta['search_year'], meta.get('imdb_id', '0'), meta.get('tvdb_id', 0), meta)
+            meta['tvmaze_id'], meta['imdb_id'], meta['tvdb_id'] = await search_tvmaze(filename, meta['search_year'], meta.get('imdb_id', '0'), meta.get('tvdb_id', 0), meta)
         else:
             meta.setdefault('tvmaze_id', '0')
         # If no imdb, search for it
         if meta.get('imdb_id', None) is None:
-            meta['imdb_id'] = await self.search_imdb(filename, meta['search_year'])
+            meta['imdb_id'] = await search_imdb(filename, meta['search_year'])
         if meta.get('imdb_info', None) is None and int(meta['imdb_id']) != 0:
-            meta['imdb_info'] = await self.get_imdb_info_api(meta['imdb_id'], meta)
+            meta['imdb_info'] = await get_imdb_info_api(meta['imdb_id'], meta)
         if meta.get('tag', None) is None:
             meta['tag'] = self.get_tag(video, meta)
         else:
@@ -812,7 +456,7 @@ class Prep():
                                     tracker_status[tracker_name]['skipped'] = True
                 if meta.get('skipping') is None and not is_dupe and tracker_name == "PTP":
                     if meta.get('imdb_info', {}) == {}:
-                        meta['imdb_info'] = self.get_imdb_info_api(meta['imdb_id'], meta)
+                        meta['imdb_info'] = get_imdb_info_api(meta['imdb_id'], meta)
                 if not meta['debug']:
                     if not tracker_status[tracker_name]['banned'] and not tracker_status[tracker_name]['skipped'] and not tracker_status[tracker_name]['dupe']:
                         console.print(f"[bold yellow]Tracker '{tracker_name}' passed all checks.")
@@ -864,40 +508,34 @@ class Prep():
         manual_frames = meta['manual_frames']
         # Take Screenshots
         if meta['is_disc'] == "BDMV":
-            if meta.get('edit', False) is False:
-                if meta.get('vapoursynth', False) is True:
-                    use_vs = True
-                else:
-                    use_vs = False
+            if not meta.get('edit', False):
+                use_vs = meta.get('vapoursynth', False)
                 try:
-                    ds = multiprocessing.Process(target=self.disc_screenshots, args=(meta, filename, bdinfo, meta['uuid'], base_dir, use_vs, meta.get('image_list', []), meta.get('ffdebug', False), None))
-                    ds.start()
-                    while ds.is_alive() is True:
-                        await asyncio.sleep(1)
-                except KeyboardInterrupt:
-                    ds.terminate()
-        elif meta['is_disc'] == "DVD":
-            if meta.get('edit', False) is False:
-                try:
-                    ds = multiprocessing.Process(target=self.dvd_screenshots, args=(meta, 0, None, None))
-                    ds.start()
-                    while ds.is_alive() is True:
-                        await asyncio.sleep(1)
-                except KeyboardInterrupt:
-                    ds.terminate()
-        else:
-            if meta.get('edit', False) is False:
-                try:
-                    s = multiprocessing.Process(
-                        target=self.screenshots,
-                        args=(videopath, filename, meta['uuid'], base_dir, meta),  # Positional arguments
-                        kwargs={'manual_frames': manual_frames}  # Keyword argument
+                    await disc_screenshots(
+                        meta, filename, bdinfo, meta['uuid'], base_dir, use_vs,
+                        meta.get('image_list', []), meta.get('ffdebug', False), None
                     )
-                    s.start()
-                    while s.is_alive() is True:
-                        await asyncio.sleep(3)
-                except KeyboardInterrupt:
-                    s.terminate()
+                except Exception as e:
+                    print(f"Error during BDMV screenshot capture: {e}")
+
+        elif meta['is_disc'] == "DVD":
+            if not meta.get('edit', False):
+                try:
+                    await dvd_screenshots(
+                        meta, 0, None, None
+                    )
+                except Exception as e:
+                    print(f"Error during DVD screenshot capture: {e}")
+
+        else:
+            if not meta.get('edit', False):
+                try:
+                    await screenshots(
+                        videopath, filename, meta['uuid'], base_dir, meta,
+                        manual_frames=manual_frames  # Pass additional kwargs directly
+                    )
+                except Exception as e:
+                    print(f"Error during generic screenshot capture: {e}")
 
         # WORK ON THIS
         meta.get('stream', False)
@@ -1369,805 +1007,6 @@ class Prep():
         return video, scene, imdb
 
     """
-    Generate Screenshots
-    """
-    def sanitize_filename(self, filename):
-        # Replace invalid characters like colons with an underscore
-        return re.sub(r'[<>:"/\\|?*]', '_', filename)
-
-    def disc_screenshots(self, meta, filename, bdinfo, folder_id, base_dir, use_vs, image_list, ffdebug, num_screens=None, force_screenshots=False):
-        if 'image_list' not in meta:
-            meta['image_list'] = []
-        existing_images = [img for img in meta['image_list'] if isinstance(img, dict) and img.get('img_url', '').startswith('http')]
-
-        if len(existing_images) >= meta.get('cutoff') and not force_screenshots:
-            console.print("[yellow]There are already at least {} images in the image list. Skipping additional screenshots.".format(meta.get('cutoff')))
-            return
-
-        if num_screens is None:
-            num_screens = self.screens
-        if num_screens == 0 or len(image_list) >= num_screens:
-            return
-
-        sanitized_filename = self.sanitize_filename(filename)
-        length = 0
-        file = None
-        frame_rate = None
-        for each in bdinfo['files']:
-            int_length = sum(int(float(x)) * 60 ** i for i, x in enumerate(reversed(each['length'].split(':'))))
-            if int_length > length:
-                length = int_length
-                for root, dirs, files in os.walk(bdinfo['path']):
-                    for name in files:
-                        if name.lower() == each['file'].lower():
-                            file = os.path.join(root, name)
-
-        if 'video' in bdinfo and bdinfo['video']:
-            fps_string = bdinfo['video'][0].get('fps', None)
-            if fps_string:
-                try:
-                    frame_rate = float(fps_string.split(' ')[0])  # Extract and convert to float
-                except ValueError:
-                    console.print("[red]Error: Unable to parse frame rate from bdinfo['video'][0]['fps']")
-
-        keyframe = 'nokey' if "VC-1" in bdinfo['video'][0]['codec'] or bdinfo['video'][0]['hdr_dv'] != "" else 'none'
-
-        os.chdir(f"{base_dir}/tmp/{folder_id}")
-        existing_screens = glob.glob(f"{sanitized_filename}-*.png")
-        total_existing = len(existing_screens) + len(existing_images)
-        if not force_screenshots:
-            num_screens = max(0, self.screens - total_existing)
-        else:
-            num_screens = num_screens
-
-        if num_screens == 0 and not force_screenshots:
-            console.print('[bold green]Reusing existing screenshots. No additional screenshots needed.')
-            return
-
-        if meta['debug'] and not force_screenshots:
-            console.print(f"[bold yellow]Saving Screens... Total needed: {self.screens}, Existing: {total_existing}, To capture: {num_screens}")
-
-        tone_map = meta.get('tone_map', False)
-        if tone_map and "HDR" in meta['hdr']:
-            hdr_tonemap = True
-        else:
-            hdr_tonemap = False
-
-        capture_tasks = []
-        capture_results = []
-        if hdr_tonemap:
-            task_limit = int(meta.get('tone_task_limit'))
-        else:
-            task_limit = int(meta.get('task_limit', os.cpu_count()))
-
-        if use_vs:
-            from src.vs import vs_screengn
-            vs_screengn(source=file, encode=None, filter_b_frames=False, num=num_screens, dir=f"{base_dir}/tmp/{folder_id}/")
-        else:
-            if meta.get('ffdebug', False):
-                loglevel = 'verbose'
-            else:
-                loglevel = 'quiet'
-
-            ss_times = self.valid_ss_time([], num_screens + 1, length, frame_rate)
-            existing_indices = {int(p.split('-')[-1].split('.')[0]) for p in existing_screens}
-            capture_tasks = [
-                (
-                    file,
-                    ss_times[i],
-                    os.path.abspath(f"{base_dir}/tmp/{folder_id}/{sanitized_filename}-{len(existing_indices) + i}.png"),
-                    keyframe,
-                    loglevel,
-                    hdr_tonemap
-                )
-                for i in range(num_screens + 1)
-            ]
-
-            with get_context("spawn").Pool(processes=min(len(capture_tasks), task_limit)) as pool:
-                try:
-                    capture_results = list(
-                        tqdm(
-                            pool.imap_unordered(self.capture_disc_task, capture_tasks),
-                            total=len(capture_tasks),
-                            desc="Capturing Screenshots",
-                            ascii=True,
-                            dynamic_ncols=False
-                        )
-                    )
-                finally:
-                    pool.close()
-                    pool.join()
-
-            if capture_results:
-                if len(capture_tasks) > num_screens:
-                    smallest = min(capture_results, key=os.path.getsize)
-                    if meta['debug']:
-                        console.print(f"[yellow]Removing smallest image: {smallest} ({os.path.getsize(smallest)} bytes)[/yellow]")
-                    os.remove(smallest)
-                    capture_results.remove(smallest)
-            optimized_results = []
-            optimize_tasks = [(result, self.config) for result in capture_results if result and os.path.exists(result)]
-            with get_context("spawn").Pool(processes=min(len(optimize_tasks), task_limit)) as pool:
-                try:
-                    optimized_results = list(
-                        tqdm(
-                            pool.imap_unordered(self.optimize_image_task, optimize_tasks),
-                            total=len(optimize_tasks),
-                            desc="Optimizing Images",
-                            ascii=True,
-                            dynamic_ncols=False
-                        )
-                    )
-                finally:
-                    pool.close()
-                    pool.join()
-
-            valid_results = []
-            remaining_retakes = []
-            for image_path in optimized_results:
-                if "Error" in image_path:
-                    console.print(f"[red]{image_path}")
-                    continue
-
-                retake = False
-                image_size = os.path.getsize(image_path)
-                if image_size <= 75000:
-                    console.print(f"[yellow]Image {image_path} is incredibly small, retaking.")
-                    retake = True
-                    time.sleep(1)
-                elif image_size <= 31000000 and self.img_host == "imgbb" and not retake:
-                    pass
-                elif image_size <= 10000000 and self.img_host in ["imgbox", "pixhost"] and not retake:
-                    pass
-                elif self.img_host in ["ptpimg", "lensdump", "ptscreens", "oeimg"] and not retake:
-                    pass
-                elif not retake:
-                    console.print("[red]Image too large for your image host, retaking.")
-                    retake = True
-                    time.sleep(1)
-
-                if retake:
-                    retry_attempts = 3
-                    for attempt in range(1, retry_attempts + 1):
-                        console.print(f"[yellow]Retaking screenshot for: {image_path} (Attempt {attempt}/{retry_attempts})[/yellow]")
-                        try:
-                            os.remove(image_path)
-                            random_time = random.uniform(0, length)
-                            self.capture_disc_task((file, random_time, image_path, keyframe, loglevel, hdr_tonemap))
-                            self.optimize_image_task((image_path, config))
-                            new_size = os.path.getsize(image_path)
-                            valid_image = False
-
-                            if new_size > 75000 and new_size <= 31000000 and self.img_host == "imgbb":
-                                console.print(f"[green]Successfully retaken screenshot for: {image_path} ({new_size} bytes)[/green]")
-                                valid_image = True
-                            elif new_size > 75000 and new_size <= 10000000 and self.img_host in ["imgbox", "pixhost"]:
-                                console.print(f"[green]Successfully retaken screenshot for: {image_path} ({new_size} bytes)[/green]")
-                                valid_image = True
-                            elif new_size > 75000 and self.img_host in ["ptpimg", "lensdump", "ptscreens", "oeimg"]:
-                                console.print(f"[green]Successfully retaken screenshot for: {image_path} ({new_size} bytes)[/green]")
-                                valid_image = True
-
-                            if valid_image:
-                                valid_results.append(image_path)
-                                break
-                            else:
-                                console.print(f"[red]Retaken image {image_path} does not meet the size requirements for {self.img_host}. Retrying...[/red]")
-                        except Exception as e:
-                            console.print(f"[red]Error retaking screenshot for {image_path}: {e}[/red]")
-                    else:
-                        console.print(f"[red]All retry attempts failed for {image_path}. Skipping.[/red]")
-                        remaining_retakes.append(image_path)
-                else:
-                    valid_results.append(image_path)
-
-            if remaining_retakes:
-                console.print(f"[red]The following images could not be retaken successfully: {remaining_retakes}[/red]")
-
-            for image_path in valid_results:
-                img_dict = {
-                    'img_url': image_path,
-                    'raw_url': image_path,
-                    'web_url': image_path
-                }
-                meta['image_list'].append(img_dict)
-
-        console.print(f"[green]Successfully captured {len(valid_results)} screenshots.")
-
-    def capture_disc_task(self, task):
-        file, ss_time, image_path, keyframe, loglevel, hdr_tonemap = task
-        try:
-            ff = ffmpeg.input(file, ss=ss_time, skip_frame=keyframe)
-
-            if hdr_tonemap:
-                ff = (
-                    ff
-                    .filter('zscale', transfer='linear')
-                    .filter('tonemap', tonemap='mobius', desat=8.0)
-                    .filter('zscale', transfer='bt709')
-                    .filter('format', 'rgb24')
-                )
-
-            command = (
-                ff
-                .output(image_path, vframes=1, pix_fmt="rgb24")
-                .overwrite_output()
-                .global_args('-loglevel', loglevel)
-            )
-
-            command.run(capture_stdout=True, capture_stderr=True)
-
-            return image_path
-        except ffmpeg.Error as e:
-            error_output = e.stderr.decode('utf-8')
-            console.print(f"[red]FFmpeg error capturing screenshot: {error_output}[/red]")
-            return None
-        except Exception as e:
-            console.print(f"[red]Error capturing screenshot: {e}[/red]")
-            return None
-
-    def dvd_screenshots(self, meta, disc_num, num_screens=None, retry_cap=None):
-        if 'image_list' not in meta:
-            meta['image_list'] = []
-        existing_images = [img for img in meta['image_list'] if isinstance(img, dict) and img.get('img_url', '').startswith('http')]
-
-        if len(existing_images) >= meta.get('cutoff') and not retry_cap:
-            console.print("[yellow]There are already at least {} images in the image list. Skipping additional screenshots.".format(meta.get('cutoff')))
-            return
-
-        if num_screens is None:
-            num_screens = self.screens - len(existing_images)
-        if num_screens == 0 or (len(meta.get('image_list', [])) >= self.screens and disc_num == 0):
-            return
-
-        if len(glob.glob(f"{meta['base_dir']}/tmp/{meta['uuid']}/{meta['discs'][disc_num]['name']}-*.png")) >= num_screens:
-            i = num_screens
-            console.print('[bold green]Reusing screenshots')
-            return
-
-        ifo_mi = MediaInfo.parse(f"{meta['discs'][disc_num]['path']}/VTS_{meta['discs'][disc_num]['main_set'][0][:2]}_0.IFO", mediainfo_options={'inform_version': '1'})
-        sar = 1
-        for track in ifo_mi.tracks:
-            if track.track_type == "Video":
-                if isinstance(track.duration, str):
-                    durations = [float(d) for d in track.duration.split(' / ')]
-                    length = max(durations) / 1000  # Use the longest duration
-                else:
-                    length = float(track.duration) / 1000  # noqa #F841 # Convert to seconds
-
-                par = float(track.pixel_aspect_ratio)
-                dar = float(track.display_aspect_ratio)
-                width = float(track.width)
-                height = float(track.height)
-                frame_rate = float(track.frame_rate)
-        if par < 1:
-            new_height = dar * height
-            sar = width / new_height
-            w_sar = 1
-            h_sar = sar
-        else:
-            sar = par
-            w_sar = sar
-            h_sar = 1
-
-        def _is_vob_good(n, loops, num_screens):
-            max_loops = 6
-            fallback_duration = 300
-            valid_tracks = []
-
-            while loops < max_loops:
-                try:
-                    vob_mi = MediaInfo.parse(
-                        f"{meta['discs'][disc_num]['path']}/VTS_{main_set[n]}",
-                        output='JSON'
-                    )
-                    vob_mi = json.loads(vob_mi)
-
-                    for track in vob_mi.get('media', {}).get('track', []):
-                        duration = float(track.get('Duration', 0))
-                        width = track.get('Width')
-                        height = track.get('Height')
-
-                        if duration > 1 and width and height:  # Minimum 1-second track
-                            valid_tracks.append({
-                                'duration': duration,
-                                'track_index': n
-                            })
-
-                    if valid_tracks:
-                        # Sort by duration, take longest track
-                        longest_track = max(valid_tracks, key=lambda x: x['duration'])
-                        return longest_track['duration'], longest_track['track_index']
-
-                except Exception as e:
-                    console.print(f"[red]Error parsing VOB {n}: {e}")
-
-                n = (n + 1) % len(main_set)
-                loops += 1
-
-            return fallback_duration, 0
-
-        main_set = meta['discs'][disc_num]['main_set'][1:] if len(meta['discs'][disc_num]['main_set']) > 1 else meta['discs'][disc_num]['main_set']
-        os.chdir(f"{meta['base_dir']}/tmp/{meta['uuid']}")
-        voblength, n = _is_vob_good(0, 0, num_screens)
-        ss_times = self.valid_ss_time([], num_screens + 1, voblength, frame_rate)
-        tasks = []
-        task_limit = int(meta.get('task_limit', os.cpu_count()))
-        for i in range(num_screens + 1):
-            image = f"{meta['base_dir']}/tmp/{meta['uuid']}/{meta['discs'][disc_num]['name']}-{i}.png"
-            input_file = f"{meta['discs'][disc_num]['path']}/VTS_{main_set[i % len(main_set)]}"
-            tasks.append((input_file, image, ss_times[i], meta, width, height, w_sar, h_sar))
-
-        with get_context("spawn").Pool(processes=min(num_screens + 1, task_limit)) as pool:
-            try:
-                results = list(tqdm(pool.imap_unordered(self.capture_dvd_screenshot, tasks), total=len(tasks), desc="Capturing Screenshots", ascii=True, dynamic_ncols=False))
-            finally:
-                pool.close()
-                pool.join()
-
-        if len(glob.glob1(f"{meta['base_dir']}/tmp/{meta['uuid']}/", f"{meta['discs'][disc_num]['name']}-*")) > num_screens:
-            smallest = None
-            smallest_size = float('inf')
-            for screens in glob.glob1(f"{meta['base_dir']}/tmp/{meta['uuid']}/", f"{meta['discs'][disc_num]['name']}-*"):
-                screen_path = os.path.join(f"{meta['base_dir']}/tmp/{meta['uuid']}/", screens)
-                try:
-                    screen_size = os.path.getsize(screen_path)
-                    if screen_size < smallest_size:
-                        smallest_size = screen_size
-                        smallest = screen_path
-                except FileNotFoundError:
-                    console.print(f"[red]File not found: {screen_path}[/red]")  # Handle potential edge cases
-                    continue
-
-            if smallest:
-                if meta['debug']:
-                    console.print(f"[yellow]Removing smallest image: {smallest} ({smallest_size} bytes)[/yellow]")
-                os.remove(smallest)
-
-        optimize_tasks = [(image, self.config) for image in results if image and os.path.exists(image)]
-
-        with get_context("spawn").Pool(processes=min(len(optimize_tasks), task_limit)) as pool:
-            try:
-                optimize_results = list(  # noqa F841
-                    tqdm(
-                        pool.imap_unordered(self.optimize_image_task, optimize_tasks),
-                        total=len(optimize_tasks),
-                        desc="Optimizing Images",
-                        ascii=True,
-                        dynamic_ncols=False
-                    )
-                )
-            finally:
-                pool.close()
-                pool.join()
-
-        valid_results = []
-        retry_attempts = 3
-
-        for image in optimize_results:
-            if "Error" in image:
-                console.print(f"[red]{image}")
-                continue
-
-            retry_cap = False
-            image_size = os.path.getsize(image)
-            if image_size <= 120000:
-                console.print(f"[yellow]Image {image} is incredibly small, retaking.")
-                retry_cap = True
-                time.sleep(1)
-
-            if retry_cap:
-                for attempt in range(1, retry_attempts + 1):
-                    console.print(f"[yellow]Retaking screenshot for: {image} (Attempt {attempt}/{retry_attempts})[/yellow]")
-                    try:
-                        os.remove(image)
-                    except Exception as e:
-                        console.print(f"[red]Failed to delete {image}: {e}[/red]")
-                        break
-
-                    image_index = int(image.rsplit('-', 1)[-1].split('.')[0])
-                    input_file = f"{meta['discs'][disc_num]['path']}/VTS_{main_set[image_index % len(main_set)]}"
-                    adjusted_time = random.uniform(0, voblength)
-
-                    try:
-                        self.capture_dvd_screenshot((input_file, image, adjusted_time, meta, width, height, w_sar, h_sar))
-                        retaken_size = os.path.getsize(image)
-
-                        if retaken_size > 75000:
-                            console.print(f"[green]Successfully retaken screenshot for: {image} ({retaken_size} bytes)[/green]")
-                            valid_results.append(image)
-                            break
-                        else:
-                            console.print(f"[red]Retaken image {image} is still too small. Retrying...[/red]")
-                    except Exception as e:
-                        console.print(f"[red]Error capturing screenshot for {input_file} at {adjusted_time}: {e}[/red]")
-
-                else:
-                    console.print(f"[red]All retry attempts failed for {image}. Skipping.[/red]")
-            else:
-                valid_results.append(image)
-
-        for image in valid_results:
-            img_dict = {
-                'img_url': image,
-                'raw_url': image,
-                'web_url': image
-            }
-            meta['image_list'].append(img_dict)
-
-        console.print(f"[green]Successfully captured {len(optimize_results)} screenshots.")
-
-    def capture_dvd_screenshot(self, task):
-        input_file, image, seek_time, meta, width, height, w_sar, h_sar = task
-
-        if os.path.exists(image):
-            console.print(f"[green]Screenshot already exists: {image}[/green]")
-            return image
-
-        try:
-            loglevel = 'verbose' if meta.get('ffdebug', False) else 'quiet'
-            media_info = MediaInfo.parse(input_file)
-            video_duration = next((track.duration for track in media_info.tracks if track.track_type == "Video"), None)
-
-            if video_duration and seek_time > video_duration:
-                seek_time = max(0, video_duration - 1)
-
-            ff = ffmpeg.input(input_file, ss=seek_time)
-            if w_sar != 1 or h_sar != 1:
-                ff = ff.filter('scale', int(round(width * w_sar)), int(round(height * h_sar)))
-
-            try:
-                ff.output(image, vframes=1, pix_fmt="rgb24").overwrite_output().global_args('-loglevel', loglevel, '-accurate_seek').run()
-            except ffmpeg._run.Error as e:
-                stderr_output = e.stderr.decode() if e.stderr else "No stderr output available"
-                console.print(f"[red]Error capturing screenshot for {input_file} at {seek_time}s: {stderr_output}[/red]")
-            if os.path.exists(image):
-                return image
-            else:
-                console.print(f"[red]Screenshot creation failed for {image}[/red]")
-                return None
-
-        except Exception as e:
-            console.print(f"[red]Error capturing screenshot for {input_file} at {seek_time}s: {e}[/red]")
-            return None
-
-    def screenshots(self, path, filename, folder_id, base_dir, meta, num_screens=None, force_screenshots=False, manual_frames=None):
-        def use_tqdm():
-            """Check if the environment supports TTY (interactive progress bar)."""
-            return sys.stdout.isatty()
-
-        if meta['debug']:
-            start_time = time.time()
-        if 'image_list' not in meta:
-            meta['image_list'] = []
-
-        existing_images = [img for img in meta['image_list'] if isinstance(img, dict) and img.get('img_url', '').startswith('http')]
-
-        if len(existing_images) >= meta.get('cutoff') and not force_screenshots:
-            console.print("[yellow]There are already at least {} images in the image list. Skipping additional screenshots.".format(meta.get('cutoff')))
-            return
-
-        if num_screens is None:
-            num_screens = self.screens - len(existing_images)
-        if num_screens <= 0:
-            return
-
-        try:
-            with open(f"{base_dir}/tmp/{folder_id}/MediaInfo.json", encoding='utf-8') as f:
-                mi = json.load(f)
-                video_track = mi['media']['track'][1]
-                length = video_track.get('Duration', mi['media']['track'][0]['Duration'])
-                width = float(video_track.get('Width'))
-                height = float(video_track.get('Height'))
-                par = float(video_track.get('PixelAspectRatio', 1))
-                dar = float(video_track.get('DisplayAspectRatio'))
-                frame_rate = float(video_track.get('FrameRate', 24.0))
-
-                if par == 1:
-                    sar = w_sar = h_sar = 1
-                elif par < 1:
-                    new_height = dar * height
-                    sar = width / new_height
-                    w_sar = 1
-                    h_sar = sar
-                else:
-                    sar = w_sar = par
-                    h_sar = 1
-                length = round(float(length))
-        except (FileNotFoundError, KeyError, ValueError) as e:
-            console.print(f"[red]Error processing MediaInfo.json: {e}")
-            return
-
-        loglevel = 'verbose' if meta.get('ffdebug', False) else 'quiet'
-        os.chdir(f"{base_dir}/tmp/{folder_id}")
-
-        if manual_frames:
-            if meta['debug']:
-                console.print(f"[yellow]Using manual frames: {manual_frames}")
-            manual_frames = [int(frame) for frame in manual_frames.split(',')]
-            ss_times = [frame / frame_rate for frame in manual_frames]
-        else:
-            ss_times = []
-
-        ss_times = self.valid_ss_time(
-            ss_times,
-            num_screens,
-            length,
-            frame_rate,
-            exclusion_zone=500
-        )
-        if meta['debug']:
-            console.print(f"[green]Final list of frames for screenshots: {ss_times}")
-
-        tone_map = meta.get('tone_map', False)
-        if tone_map and "HDR" in meta['hdr']:
-            hdr_tonemap = True
-        else:
-            hdr_tonemap = False
-
-        capture_tasks = []
-        capture_results = []
-        if hdr_tonemap:
-            task_limit = int(meta.get('tone_task_limit'))
-        else:
-            task_limit = int(meta.get('task_limit', os.cpu_count()))
-
-        existing_images = 0
-        for i in range(num_screens):
-            image_path = os.path.abspath(f"{base_dir}/tmp/{folder_id}/{filename}-{i}.png")
-            if os.path.exists(image_path) and not meta.get('retake', False):
-                existing_images += 1
-
-        if existing_images == num_screens and not meta.get('retake', False):
-            console.print("[yellow]The correct number of screenshots already exists. Skipping capture process.")
-        else:
-            for i in range(num_screens + 1):
-                image_path = os.path.abspath(f"{base_dir}/tmp/{folder_id}/{filename}-{i}.png")
-                if not os.path.exists(image_path) or meta.get('retake', False):
-                    capture_tasks.append((path, ss_times[i], image_path, width, height, w_sar, h_sar, loglevel, hdr_tonemap))
-                elif meta['debug']:
-                    console.print(f"[yellow]Skipping existing screenshot: {image_path}")
-
-            if not capture_tasks:
-                console.print("[yellow]All screenshots already exist. Skipping capture process.")
-            else:
-                if use_tqdm():
-                    with tqdm(total=len(capture_tasks), desc="Capturing Screenshots", ascii=True, dynamic_ncols=False) as pbar:
-                        with get_context("spawn").Pool(processes=min(len(capture_tasks), task_limit)) as pool:
-                            try:
-                                for result in pool.imap_unordered(self.capture_screenshot, capture_tasks):
-                                    if isinstance(result, str) and result.startswith("Error:"):
-                                        console.print(f"[red]Capture Error: {result}")
-                                    else:
-                                        capture_results.append(result)
-                                    pbar.update(1)
-                            finally:
-                                pool.close()
-                                pool.join()
-                else:
-                    console.print("[blue]Non-TTY environment detected. Progress bar disabled.")
-                    with get_context("spawn").Pool(processes=min(len(capture_tasks), task_limit)) as pool:
-                        try:
-                            for i, result in enumerate(pool.imap_unordered(self.capture_screenshot, capture_tasks), 1):
-                                capture_results.append(result)
-                                console.print(f"Processed {i}/{len(capture_tasks)} screenshots")
-                        finally:
-                            pool.close()
-                            pool.join()
-
-                if capture_results and (len(capture_results) + existing_images) > num_screens and not force_screenshots:
-                    smallest = min(capture_results, key=os.path.getsize)
-                    if meta['debug']:
-                        console.print(f"[yellow]Removing smallest image: {smallest} ({os.path.getsize(smallest)} bytes)[/yellow]")
-                    os.remove(smallest)
-                    capture_results.remove(smallest)
-
-        optimize_tasks = [(result, self.config) for result in capture_results if "Error" not in result]
-        optimize_results = []
-        if optimize_tasks:
-            if use_tqdm():
-                with tqdm(total=len(optimize_tasks), desc="Optimizing Images", ascii=True, dynamic_ncols=False) as pbar:
-                    with get_context("spawn").Pool(processes=min(len(optimize_tasks), task_limit)) as pool:
-                        try:
-                            for result in pool.imap_unordered(self.optimize_image_task, optimize_tasks):
-                                optimize_results.append(result)
-                                pbar.update(1)
-                        finally:
-                            pool.close()
-                            pool.join()
-            else:
-                with get_context("spawn").Pool(processes=min(len(optimize_tasks), task_limit)) as pool:
-                    try:
-                        for i, result in enumerate(pool.imap_unordered(self.optimize_image_task, optimize_tasks), 1):
-                            optimize_results.append(result)
-                            console.print(f"Optimized {i}/{len(optimize_tasks)} images")
-                    finally:
-                        pool.close()
-                        pool.join()
-
-        valid_results = []
-        remaining_retakes = []
-        for image_path in optimize_results:
-            if "Error" in image_path:
-                console.print(f"[red]{image_path}")
-                continue
-
-            retake = False
-            image_size = os.path.getsize(image_path)
-            if not manual_frames:
-                if image_size <= 75000:
-                    console.print(f"[yellow]Image {image_path} is incredibly small, retaking.")
-                    retake = True
-                    time.sleep(1)
-                elif image_size <= 31000000 and self.img_host == "imgbb" and not retake:
-                    pass
-                elif image_size <= 10000000 and self.img_host in ["imgbox", "pixhost"] and not retake:
-                    pass
-                elif self.img_host in ["ptpimg", "lensdump", "ptscreens", "oeimg"] and not retake:
-                    pass
-                elif not retake:
-                    console.print("[red]Image too large for your image host, retaking.")
-                    retake = True
-                    time.sleep(1)
-
-            if retake:
-                retry_attempts = 3
-                for attempt in range(1, retry_attempts + 1):
-                    console.print(f"[yellow]Retaking screenshot for: {image_path} (Attempt {attempt}/{retry_attempts})[/yellow]")
-                    try:
-                        os.remove(image_path)
-                        random_time = random.uniform(0, length)
-                        self.capture_screenshot((path, random_time, image_path, width, height, w_sar, h_sar, loglevel, hdr_tonemap))
-                        self.optimize_image_task((image_path, config))
-                        new_size = os.path.getsize(image_path)
-                        valid_image = False
-
-                        if new_size > 75000 and new_size <= 31000000 and self.img_host == "imgbb":
-                            console.print(f"[green]Successfully retaken screenshot for: {image_path} ({new_size} bytes)[/green]")
-                            valid_image = True
-                        elif new_size > 75000 and new_size <= 10000000 and self.img_host in ["imgbox", "pixhost"]:
-                            console.print(f"[green]Successfully retaken screenshot for: {image_path} ({new_size} bytes)[/green]")
-                            valid_image = True
-                        elif new_size > 75000 and self.img_host in ["ptpimg", "lensdump", "ptscreens", "oeimg"]:
-                            console.print(f"[green]Successfully retaken screenshot for: {image_path} ({new_size} bytes)[/green]")
-                            valid_image = True
-
-                        if valid_image:
-                            valid_results.append(image_path)
-                            break
-                        else:
-                            console.print(f"[red]Retaken image {image_path} does not meet the size requirements for {self.img_host}. Retrying...[/red]")
-                    except Exception as e:
-                        console.print(f"[red]Error retaking screenshot for {image_path}: {e}[/red]")
-                else:
-                    console.print(f"[red]All retry attempts failed for {image_path}. Skipping.[/red]")
-                    remaining_retakes.append(image_path)
-            else:
-                valid_results.append(image_path)
-
-        if remaining_retakes:
-            console.print(f"[red]The following images could not be retaken successfully: {remaining_retakes}[/red]")
-
-        for image_path in valid_results:
-            img_dict = {
-                'img_url': image_path,
-                'raw_url': image_path,
-                'web_url': image_path
-            }
-            meta['image_list'].append(img_dict)
-
-        console.print(f"[green]Successfully captured {len(valid_results)} screenshots.")
-
-        if meta['debug']:
-            finish_time = time.time()
-            console.print(f"Screenshots processed in {finish_time - start_time:.4f} seconds")
-
-    def valid_ss_time(self, ss_times, num_screens, length, frame_rate, exclusion_zone=None):
-        total_screens = num_screens + 1
-
-        if exclusion_zone is None:
-            exclusion_zone = max(length / (3 * total_screens), length / 15)
-
-        result_times = ss_times.copy()
-        section_size = (round(4 * length / 5) - round(length / 5)) / total_screens * 1.3
-        section_starts = [round(length / 5) + i * (section_size * 0.9) for i in range(total_screens)]
-
-        for section_index in range(total_screens):
-            valid_time = False
-            attempts = 0
-            start_frame = round(section_starts[section_index] * frame_rate)
-            end_frame = round((section_starts[section_index] + section_size) * frame_rate)
-
-            while not valid_time and attempts < 50:
-                attempts += 1
-                frame = random.randint(start_frame, end_frame)
-                time = frame / frame_rate
-
-                if all(abs(frame - existing_time * frame_rate) > exclusion_zone * frame_rate for existing_time in result_times):
-                    result_times.append(time)
-                    valid_time = True
-
-            if not valid_time:
-                midpoint_frame = (start_frame + end_frame) // 2
-                result_times.append(midpoint_frame / frame_rate)
-
-        result_times = sorted(result_times)
-
-        return result_times
-
-    def capture_screenshot(self, args):
-        path, ss_time, image_path, width, height, w_sar, h_sar, loglevel, hdr_tonemap = args
-        try:
-            if width <= 0 or height <= 0:
-                return "Error: Invalid width or height for scaling"
-
-            if ss_time < 0:
-                return f"Error: Invalid timestamp {ss_time}"
-
-            ff = ffmpeg.input(path, ss=ss_time)
-            if w_sar != 1 or h_sar != 1:
-                ff = ff.filter('scale', int(round(width * w_sar)), int(round(height * h_sar)))
-
-            if hdr_tonemap:
-                ff = (
-                    ff
-                    .filter('zscale', transfer='linear')
-                    .filter('tonemap', tonemap='mobius', desat=8.0)
-                    .filter('zscale', transfer='bt709')
-                    .filter('format', 'rgb24')
-                )
-
-            command = (
-                ff
-                .output(
-                    image_path,
-                    vframes=1,
-                    pix_fmt="rgb24"
-                )
-                .overwrite_output()
-                .global_args('-loglevel', loglevel)
-            )
-
-            try:
-                command.run(capture_stdout=True, capture_stderr=True)
-            except ffmpeg.Error as e:
-                error_output = e.stderr.decode('utf-8')
-                return f"Error: {error_output}"
-
-            if not os.path.exists(image_path) or os.path.getsize(image_path) == 0:
-                return f"Error: Screenshot not generated or is empty at {image_path}"
-
-            return image_path
-        except Exception as e:
-            return f"Error: {str(e)}"
-
-    def optimize_image_task(self, args):
-        image, config = args
-        try:
-            # Extract shared_seedbox and optimize_images from config
-            optimize_images = config['DEFAULT'].get('optimize_images', True)
-            shared_seedbox = config['DEFAULT'].get('shared_seedbox', True)
-
-            if optimize_images:
-                if shared_seedbox:
-                    # Limit the number of threads for oxipng
-                    num_cores = multiprocessing.cpu_count()
-                    max_threads = num_cores // 2
-                    os.environ['RAYON_NUM_THREADS'] = str(max_threads)
-
-                if os.path.exists(image):
-                    pyver = platform.python_version_tuple()
-                    if int(pyver[0]) == 3 and int(pyver[1]) >= 7:
-                        import oxipng
-                    if os.path.getsize(image) >= 16000000:
-                        oxipng.optimize(image, level=6)
-                    else:
-                        oxipng.optimize(image, level=2)
-            return image  # Return image path if successful
-        except (KeyboardInterrupt, Exception) as e:
-            return f"Error: {e}"  # Return error message
-
-    """
     Get type and category
     """
 
@@ -2224,7 +1063,7 @@ class Prep():
             meta['tmdb'] = info['tv_results'][0]['id']
             meta['original_language'] = info['tv_results'][0].get('original_language')
         else:
-            imdb_info = await self.get_imdb_info_api(imdb_id.replace('tt', ''), meta)
+            imdb_info = await get_imdb_info_api(imdb_id.replace('tt', ''), meta)
             title = imdb_info.get("title")
             if title is None:
                 title = filename
@@ -2334,7 +1173,7 @@ class Prep():
             except Exception:
                 console.print('[yellow]Unable to grab videos from TMDb.')
 
-            meta['aka'], original_language = await self.get_imdb_aka_api(meta['imdb_id'], meta)
+            meta['aka'], original_language = await get_imdb_aka_api(meta['imdb_id'], meta)
             if original_language is not None:
                 meta['original_language'] = original_language
             else:
@@ -2388,7 +1227,7 @@ class Prep():
                 console.print('[yellow]Unable to grab videos from TMDb.')
 
             # meta['aka'] = f" AKA {response['original_name']}"
-            meta['aka'], original_language = await self.get_imdb_aka_api(meta['imdb_id'], meta)
+            meta['aka'], original_language = await get_imdb_aka_api(meta['imdb_id'], meta)
             if original_language is not None:
                 meta['original_language'] = original_language
             else:
@@ -3333,374 +2172,6 @@ class Prep():
             base_torrent.private = True
             Torrent.copy(base_torrent).write(f"{base_dir}/tmp/{uuid}/BASE.torrent", overwrite=True)
 
-    """
-    Upload Screenshots
-    """
-    def upload_image_task(self, args):
-        image, img_host, config, meta = args
-        try:
-            timeout = 60  # Default timeout
-            img_url, raw_url, web_url = None, None, None
-
-            if img_host == "imgbox":
-                try:
-                    # Call the asynchronous imgbox_upload function
-                    loop = asyncio.get_event_loop()
-                    image_list = loop.run_until_complete(
-                        self.imgbox_upload(os.getcwd(), [image], meta, return_dict={})
-                    )
-                    if image_list and all(
-                        'img_url' in img and 'raw_url' in img and 'web_url' in img for img in image_list
-                    ):
-                        img_url = image_list[0]['img_url']
-                        raw_url = image_list[0]['raw_url']
-                        web_url = image_list[0]['web_url']
-                    else:
-                        return {
-                            'status': 'failed',
-                            'reason': "Imgbox upload failed. No valid URLs returned."
-                        }
-                except Exception as e:
-                    return {
-                        'status': 'failed',
-                        'reason': f"Error during Imgbox upload: {str(e)}"
-                    }
-
-            elif img_host == "ptpimg":
-                payload = {
-                    'format': 'json',
-                    'api_key': config['DEFAULT']['ptpimg_api']
-                }
-                files = [('file-upload[0]', open(image, 'rb'))]
-                headers = {'referer': 'https://ptpimg.me/index.php'}
-                response = requests.post(
-                    "https://ptpimg.me/upload.php", headers=headers, data=payload, files=files, timeout=timeout
-                )
-                response_data = response.json()
-                if response_data:
-                    code = response_data[0]['code']
-                    ext = response_data[0]['ext']
-                    img_url = f"https://ptpimg.me/{code}.{ext}"
-                    raw_url = img_url
-                    web_url = img_url
-
-            elif img_host == "imgbb":
-                url = "https://api.imgbb.com/1/upload"
-                try:
-                    with open(image, "rb") as img_file:
-                        encoded_image = base64.b64encode(img_file.read()).decode('utf8')
-
-                    data = {
-                        'key': config['DEFAULT']['imgbb_api'],
-                        'image': encoded_image,
-                    }
-
-                    response = requests.post(url, data=data, timeout=timeout)
-
-                    if meta['debug']:
-                        console.print(f"[yellow]Response status code: {response.status_code}")
-                        console.print(f"[yellow]Response content: {response.content.decode('utf-8')}")
-
-                    response_data = response.json()
-                    if response.status_code != 200 or not response_data.get('success'):
-                        console.print("[yellow]imgbb failed, trying next image host")
-                        return {'status': 'failed', 'reason': 'imgbb upload failed'}
-
-                    img_url = response_data['data'].get('medium', {}).get('url') or response_data['data']['thumb']['url']
-                    raw_url = response_data['data']['image']['url']
-                    web_url = response_data['data']['url_viewer']
-
-                    if meta['debug']:
-                        console.print(f"[green]Image URLs: img_url={img_url}, raw_url={raw_url}, web_url={web_url}")
-
-                    return {'status': 'success', 'img_url': img_url, 'raw_url': raw_url, 'web_url': web_url}
-
-                except requests.exceptions.Timeout:
-                    console.print("[red]Request timed out. The server took too long to respond.")
-                    return {'status': 'failed', 'reason': 'Request timed out'}
-
-                except ValueError as e:  # JSON decoding error
-                    console.print(f"[red]Invalid JSON response: {e}")
-                    return {'status': 'failed', 'reason': 'Invalid JSON response'}
-
-                except requests.exceptions.RequestException as e:
-                    console.print(f"[red]Request failed with error: {e}")
-                    return {'status': 'failed', 'reason': str(e)}
-
-            elif img_host == "ptscreens":
-                url = "https://ptscreens.com/api/1/upload"
-                try:
-                    files = {
-                        'source': ('file-upload[0]', open(image, 'rb')),
-                    }
-                    headers = {
-                        'X-API-Key': config['DEFAULT']['ptscreens_api']
-                    }
-                    response = requests.post(url, headers=headers, files=files, timeout=timeout)
-                    if meta['debug']:
-                        console.print(f"[yellow]Response status code: {response.status_code}")
-                        console.print(f"[yellow]Response content: {response.content.decode('utf-8')}")
-
-                    response_data = response.json()
-                    if response_data.get('status_code') != 200:
-                        console.print("[yellow]ptscreens failed, trying next image host")
-                        return {'status': 'failed', 'reason': 'ptscreens upload failed'}
-
-                    img_url = response_data['image']['medium']['url']
-                    raw_url = response_data['image']['url']
-                    web_url = response_data['image']['url_viewer']
-                    if meta['debug']:
-                        console.print(f"[green]Image URLs: img_url={img_url}, raw_url={raw_url}, web_url={web_url}")
-
-                except requests.exceptions.Timeout:
-                    console.print("[red]Request timed out. The server took too long to respond.")
-                    return {'status': 'failed', 'reason': 'Request timed out'}
-                except requests.exceptions.RequestException as e:
-                    console.print(f"[red]Request failed with error: {e}")
-                    return {'status': 'failed', 'reason': str(e)}
-
-            elif img_host == "oeimg":
-                url = "https://imgoe.download/api/1/upload"
-                try:
-                    data = {
-                        'image': base64.b64encode(open(image, "rb").read()).decode('utf8')
-                    }
-                    headers = {
-                        'X-API-Key': self.config['DEFAULT']['oeimg_api'],
-                    }
-                    response = requests.post(url, data=data, headers=headers, timeout=timeout)
-                    if meta['debug']:
-                        console.print(f"[yellow]Response status code: {response.status_code}")
-                        console.print(f"[yellow]Response content: {response.content.decode('utf-8')}")
-
-                    response_data = response.json()
-                    if response.status_code != 200 or not response_data.get('success'):
-                        console.print("[yellow]OEimg failed, trying next image host")
-                        return {'status': 'failed', 'reason': 'OEimg upload failed'}
-
-                    img_url = response_data['data']['image']['url']
-                    raw_url = response_data['data']['image']['url']
-                    web_url = response_data['data']['url_viewer']
-                    if meta['debug']:
-                        console.print(f"[green]Image URLs: img_url={img_url}, raw_url={raw_url}, web_url={web_url}")
-
-                except requests.exceptions.Timeout:
-                    console.print("[red]Request timed out. The server took too long to respond.")
-                    return {'status': 'failed', 'reason': 'Request timed out'}
-                except requests.exceptions.RequestException as e:
-                    console.print(f"[red]Request failed with error: {e}")
-                    return {'status': 'failed', 'reason': str(e)}
-
-            elif img_host == "pixhost":
-                url = "https://api.pixhost.to/images"
-                data = {
-                    'content_type': '0',
-                    'max_th_size': 350
-                }
-                files = {
-                    'img': ('file-upload[0]', open(image, 'rb'))
-                }
-                response = requests.post(url, data=data, files=files, timeout=timeout)
-                response_data = response.json()
-                if response.status_code == 200:
-                    raw_url = response_data['th_url'].replace('https://t', 'https://img').replace('/thumbs/', '/images/')
-                    img_url = response_data['th_url']
-                    web_url = response_data['show_url']
-
-            elif img_host == "lensdump":
-                url = "https://lensdump.com/api/1/upload"
-                data = {
-                    'image': base64.b64encode(open(image, "rb").read()).decode('utf8')
-                }
-                headers = {
-                    'X-API-Key': config['DEFAULT']['lensdump_api']
-                }
-                response = requests.post(url, data=data, headers=headers, timeout=timeout)
-                response_data = response.json()
-                if response_data.get('status_code') == 200:
-                    img_url = response_data['data']['image']['url']
-                    raw_url = response_data['data']['image']['url']
-                    web_url = response_data['data']['url_viewer']
-
-            if img_url and raw_url and web_url:
-                return {
-                    'status': 'success',
-                    'img_url': img_url,
-                    'raw_url': raw_url,
-                    'web_url': web_url,
-                    'local_file_path': image
-                }
-            else:
-                return {
-                    'status': 'failed',
-                    'reason': f"Failed to upload image to {img_host}. No URLs received."
-                }
-
-        except Exception as e:
-            return {
-                'status': 'failed',
-                'reason': str(e)
-            }
-
-    def upload_screens(self, meta, screens, img_host_num, i, total_screens, custom_img_list, return_dict, retry_mode=False, max_retries=3):
-        def use_tqdm():
-            """Check if the environment supports TTY (interactive progress bar)."""
-            return sys.stdout.isatty()
-
-        if meta['debug']:
-            upload_start_time = time.time()
-
-        import nest_asyncio
-        nest_asyncio.apply()
-        os.chdir(f"{meta['base_dir']}/tmp/{meta['uuid']}")
-        initial_img_host = self.config['DEFAULT'][f'img_host_{img_host_num}']
-        img_host = meta['imghost']
-        using_custom_img_list = isinstance(custom_img_list, list) and bool(custom_img_list)
-
-        if 'image_sizes' not in meta:
-            meta['image_sizes'] = {}
-
-        if using_custom_img_list:
-            image_glob = custom_img_list
-            existing_images = []
-            existing_count = 0
-        else:
-            image_glob = glob.glob("*.png")
-            if 'POSTER.png' in image_glob:
-                image_glob.remove('POSTER.png')
-            image_glob = list(set(image_glob))
-            if meta['debug']:
-                console.print("image globs:", image_glob)
-
-            existing_images = [img for img in meta['image_list'] if img.get('img_url') and img.get('web_url')]
-            existing_count = len(existing_images)
-
-        if not retry_mode:
-            images_needed = max(0, total_screens - existing_count)
-        else:
-            images_needed = total_screens
-
-        if existing_count >= total_screens and not retry_mode and img_host == initial_img_host and not using_custom_img_list:
-            console.print(f"[yellow]Skipping upload because enough images are already uploaded to {img_host}. Existing images: {existing_count}, Required: {total_screens}")
-            return meta['image_list'], total_screens
-
-        upload_tasks = [(image, img_host, self.config, meta) for image in image_glob[:images_needed]]
-
-        host_limits = {
-            "oeimg": 6,
-            "ptscreens": 1,
-            "lensdump": 1,
-        }
-        default_pool_size = int(meta.get('task_limit', os.cpu_count()))
-        pool_size = host_limits.get(img_host, default_pool_size)
-
-        try:
-            with get_context("spawn").Pool(processes=max(1, min(len(upload_tasks), pool_size))) as pool:
-                if use_tqdm():
-                    try:
-                        results = list(
-                            tqdm(
-                                pool.imap_unordered(self.upload_image_task, upload_tasks),
-                                total=len(upload_tasks),
-                                desc=f"Uploading Images to {img_host}",
-                                ascii=True,
-                                dynamic_ncols=False
-                            )
-                        )
-                    finally:
-                        pool.close()
-                        pool.join()
-                else:
-                    console.print(f"[blue]Non-TTY environment detected. Progress bar disabled. Uploading images to {img_host}.")
-                    results = []
-                    for i, result in enumerate(pool.imap_unordered(self.upload_image_task, upload_tasks), 1):
-                        results.append(result)
-                        console.print(f"Uploaded {i}/{len(upload_tasks)} images to {img_host}")
-        except KeyboardInterrupt:
-            console.print("[red]Upload process interrupted by user. Exiting...")
-            pool.terminate()
-            pool.join()
-            return meta['image_list'], len(meta['image_list'])
-
-        successfully_uploaded = []
-        for result in results:
-            if result['status'] == 'success':
-                successfully_uploaded.append(result)
-            else:
-                console.print(f"[yellow]Failed to upload: {result.get('reason', 'Unknown error')}")
-
-        if len(successfully_uploaded) < meta.get('cutoff') and not retry_mode and img_host == initial_img_host and not using_custom_img_list:
-            img_host_num += 1
-            if f'img_host_{img_host_num}' in self.config['DEFAULT']:
-                meta['imghost'] = self.config['DEFAULT'][f'img_host_{img_host_num}']
-                console.print(f"[cyan]Switching to the next image host: {meta['imghost']}")
-                return self.upload_screens(meta, screens, img_host_num, i, total_screens, custom_img_list, return_dict, retry_mode=True)
-            else:
-                console.print("[red]No more image hosts available. Aborting upload process.")
-                return meta['image_list'], len(meta['image_list'])
-
-        new_images = []
-        for upload in successfully_uploaded:
-            raw_url = upload['raw_url']
-            new_image = {
-                'img_url': upload['img_url'],
-                'raw_url': raw_url,
-                'web_url': upload['web_url']
-            }
-            new_images.append(new_image)
-            if not using_custom_img_list and raw_url not in {img['raw_url'] for img in meta['image_list']}:
-                if meta['debug']:
-                    console.print(f"[blue]Adding {raw_url} to image_list")
-                meta['image_list'].append(new_image)
-                local_file_path = upload.get('local_file_path')
-                if local_file_path:
-                    image_size = os.path.getsize(local_file_path)
-                    meta['image_sizes'][raw_url] = image_size
-
-        console.print(f"[green]Successfully uploaded {len(new_images)} images.")
-        if meta['debug']:
-            upload_finish_time = time.time()
-            print(f"Screenshot uploads processed in {upload_finish_time - upload_start_time:.4f} seconds")
-
-        if using_custom_img_list:
-            return new_images, len(new_images)
-
-        return meta['image_list'], len(successfully_uploaded)
-
-    async def imgbox_upload(self, chdir, image_glob, meta, return_dict):
-        try:
-            os.chdir(chdir)
-            image_list = []
-
-            async with pyimgbox.Gallery(thumb_width=350, square_thumbs=False) as gallery:
-                for image in image_glob:
-                    try:
-                        async for submission in gallery.add([image]):
-                            if not submission['success']:
-                                console.print(f"[red]Error uploading to imgbox: [yellow]{submission['error']}[/yellow][/red]")
-                            else:
-                                web_url = submission.get('web_url')
-                                img_url = submission.get('thumbnail_url')
-                                raw_url = submission.get('image_url')
-                                if web_url and img_url and raw_url:
-                                    image_dict = {
-                                        'web_url': web_url,
-                                        'img_url': img_url,
-                                        'raw_url': raw_url
-                                    }
-                                    image_list.append(image_dict)
-                                else:
-                                    console.print(f"[red]Incomplete URLs received for image: {image}")
-                    except Exception as e:
-                        console.print(f"[red]Error during upload for {image}: {str(e)}")
-
-            return_dict['image_list'] = image_list
-            return image_list
-
-        except Exception as e:
-            console.print(f"[red]An error occurred while uploading images to imgbox: {str(e)}")
-            return []
-
     async def get_name(self, meta):
         type = meta.get('type', "").upper()
         title = meta.get('title', "")
@@ -4296,158 +2767,6 @@ class Prep():
                         meta[key] = value.get(key)
         return meta
 
-    async def package(self, meta):
-        if meta['tag'] == "":
-            tag = ""
-        else:
-            tag = f" / {meta['tag'][1:]}"
-        if meta['is_disc'] == "DVD":
-            res = meta['source']
-        else:
-            res = meta['resolution']
-
-        with open(f"{meta['base_dir']}/tmp/{meta['uuid']}/GENERIC_INFO.txt", 'w', encoding="utf-8") as generic:
-            generic.write(f"Name: {meta['name']}\n\n")
-            generic.write(f"Overview: {meta['overview']}\n\n")
-            generic.write(f"{res} / {meta['type']}{tag}\n\n")
-            generic.write(f"Category: {meta['category']}\n")
-            generic.write(f"TMDB: https://www.themoviedb.org/{meta['category'].lower()}/{meta['tmdb']}\n")
-            if meta['imdb_id'] != "0":
-                generic.write(f"IMDb: https://www.imdb.com/title/tt{meta['imdb_id']}\n")
-            if meta['tvdb_id'] != "0":
-                generic.write(f"TVDB: https://www.thetvdb.com/?id={meta['tvdb_id']}&tab=series\n")
-            if "tvmaze_id" in meta and meta['tvmaze_id'] != "0":
-                generic.write(f"TVMaze: https://www.tvmaze.com/shows/{meta['tvmaze_id']}\n")
-            poster_img = f"{meta['base_dir']}/tmp/{meta['uuid']}/POSTER.png"
-            if meta.get('poster', None) not in ['', None] and not os.path.exists(poster_img):
-                if meta.get('rehosted_poster', None) is None:
-                    r = requests.get(meta['poster'], stream=True)
-                    if r.status_code == 200:
-                        console.print("[bold yellow]Rehosting Poster")
-                        r.raw.decode_content = True
-                        with open(poster_img, 'wb') as f:
-                            shutil.copyfileobj(r.raw, f)
-                        poster, dummy = self.upload_screens(meta, 1, 1, 0, 1, [poster_img], {})
-                        poster = poster[0]
-                        generic.write(f"TMDB Poster: {poster.get('raw_url', poster.get('img_url'))}\n")
-                        meta['rehosted_poster'] = poster.get('raw_url', poster.get('img_url'))
-                        with open(f"{meta['base_dir']}/tmp/{meta['uuid']}/meta.json", 'w') as metafile:
-                            json.dump(meta, metafile, indent=4)
-                            metafile.close()
-                    else:
-                        console.print("[bold yellow]Poster could not be retrieved")
-            elif os.path.exists(poster_img) and meta.get('rehosted_poster') is not None:
-                generic.write(f"TMDB Poster: {meta.get('rehosted_poster')}\n")
-            if len(meta['image_list']) > 0:
-                generic.write("\nImage Webpage:\n")
-                for each in meta['image_list']:
-                    generic.write(f"{each['web_url']}\n")
-                generic.write("\nThumbnail Image:\n")
-                for each in meta['image_list']:
-                    generic.write(f"{each['img_url']}\n")
-        title = re.sub(r"[^0-9a-zA-Z\[\\]]+", "", meta['title'])
-        archive = f"{meta['base_dir']}/tmp/{meta['uuid']}/{title}"
-        torrent_files = glob.glob1(f"{meta['base_dir']}/tmp/{meta['uuid']}", "*.torrent")
-        if isinstance(torrent_files, list) and len(torrent_files) > 1:
-            for each in torrent_files:
-                if not each.startswith(('BASE', '[RAND')):
-                    os.remove(os.path.abspath(f"{meta['base_dir']}/tmp/{meta['uuid']}/{each}"))
-        try:
-            if os.path.exists(f"{meta['base_dir']}/tmp/{meta['uuid']}/BASE.torrent"):
-                base_torrent = Torrent.read(f"{meta['base_dir']}/tmp/{meta['uuid']}/BASE.torrent")
-                manual_name = re.sub(r"[^0-9a-zA-Z\[\]\'\-]+", ".", os.path.basename(meta['path']))
-                Torrent.copy(base_torrent).write(f"{meta['base_dir']}/tmp/{meta['uuid']}/{manual_name}.torrent", overwrite=True)
-                # shutil.copy(os.path.abspath(f"{meta['base_dir']}/tmp/{meta['uuid']}/BASE.torrent"), os.path.abspath(f"{meta['base_dir']}/tmp/{meta['uuid']}/{meta['name'].replace(' ', '.')}.torrent").replace(' ', '.'))
-            filebrowser = self.config['TRACKERS'].get('MANUAL', {}).get('filebrowser', None)
-            shutil.make_archive(archive, 'tar', f"{meta['base_dir']}/tmp/{meta['uuid']}")
-            if filebrowser is not None:
-                url = '/'.join(s.strip('/') for s in (filebrowser, f"/tmp/{meta['uuid']}"))
-                url = urllib.parse.quote(url, safe="https://")
-            else:
-                files = {
-                    "files[]": (f"{meta['title']}.tar", open(f"{archive}.tar", 'rb'))
-                }
-                response = requests.post("https://uguu.se/upload.php", files=files).json()
-                if meta['debug']:
-                    console.print(f"[cyan]{response}")
-                url = response['files'][0]['url']
-            return url
-        except Exception:
-            return False
-        return
-
-    async def get_imdb_aka_api(self, imdb_id, meta):
-        if imdb_id == "0":
-            return "", None
-        if not imdb_id.startswith("tt"):
-            imdb_id = f"tt{imdb_id}"
-        url = "https://api.graphql.imdb.com/"
-        query = {
-            "query": f"""
-                query {{
-                    title(id: "{imdb_id}") {{
-                        id
-                        titleText {{
-                            text
-                            isOriginalTitle
-                        }}
-                        originalTitleText {{
-                            text
-                        }}
-                        countriesOfOrigin {{
-                            countries {{
-                                id
-                            }}
-                        }}
-                    }}
-                }}
-            """
-        }
-
-        headers = {
-            "Content-Type": "application/json",
-        }
-
-        response = requests.post(url, headers=headers, json=query)
-        data = response.json()
-
-        # Check if `data` and `title` exist
-        title_data = data.get("data", {}).get("title")
-        if title_data is None:
-            console.print("Title data is missing from response")
-            return "", None
-
-        # Extract relevant fields from the response
-        aka = title_data.get("originalTitleText", {}).get("text", "")
-        is_original = title_data.get("titleText", {}).get("isOriginalTitle", False)
-        if meta.get('manual_language'):
-            original_language = meta.get('manual_language')
-        else:
-            original_language = None
-
-        if not is_original and aka:
-            aka = f" AKA {aka}"
-
-        return aka, original_language
-
-    async def get_imdb_aka(self, imdb_id):
-        if imdb_id == "0":
-            return "", None
-        if not imdb_id.startswith("tt"):
-            imdb_id = f"tt{imdb_id}"
-        ia = Cinemagoer()
-        result = ia.get_movie(imdb_id.replace('tt', ''))
-        original_language = result.get('language codes')
-        if isinstance(original_language, list):
-            if len(original_language) > 1:
-                original_language = None
-            elif len(original_language) == 1:
-                original_language = original_language[0]
-        aka = result.get('original title', result.get('localized title', "")).replace(' - IMDb', '').replace('\u00ae', '')
-        if aka != "":
-            aka = f" AKA {aka}"
-        return aka, original_language
-
     async def get_dvd_size(self, discs, manual_dvds):
         sizes = []
         dvd_sizes = []
@@ -4500,321 +2819,3 @@ class Prep():
         else:
             console.print(f"[yellow]Unable to map the date ([bold yellow]{str(date)}[/bold yellow]) to a Season/Episode number")
         return season, episode
-
-    def safe_get(self, data, path, default=None):
-        for key in path:
-            if isinstance(data, dict):
-                data = data.get(key, default)
-            else:
-                return default
-        return data
-
-    async def get_imdb_info_api(self, imdbID, meta):
-        imdb_info = {
-            'title': meta['title'],
-            'year': meta['year'],
-            'aka': '',
-            'type': None,
-            'runtime': meta.get('runtime', '60'),
-            'cover': meta.get('poster'),
-        }
-        if len(meta.get('tmdb_directors', [])) >= 1:
-            imdb_info['directors'] = meta['tmdb_directors']
-
-        if imdbID == "0":
-            return imdb_info
-        else:
-            try:
-                if not imdbID.startswith("tt"):
-                    imdbIDtt = f"tt{imdbID}"
-                else:
-                    imdbIDtt = imdbID
-            except Exception:
-                return imdb_info
-            query = {
-                "query": f"""
-                query GetTitleInfo {{
-                  title(id: "{imdbIDtt}") {{
-                    id
-                    titleText {{
-                      text
-                      isOriginalTitle
-                    }}
-                    originalTitleText {{
-                      text
-                    }}
-                    releaseYear {{
-                      year
-                    }}
-                    titleType {{
-                      id
-                    }}
-                    plot {{
-                      plotText {{
-                        plainText
-                      }}
-                    }}
-                    ratingsSummary {{
-                      aggregateRating
-                      voteCount
-                    }}
-                    primaryImage {{
-                      url
-                    }}
-                    runtime {{
-                      displayableProperty {{
-                        value {{
-                          plainText
-                        }}
-                      }}
-                      seconds
-                    }}
-                    titleGenres {{
-                      genres {{
-                        genre {{
-                          text
-                        }}
-                      }}
-                    }}
-                    principalCredits {{
-                      category {{
-                        text
-                        id
-                      }}
-                      credits {{
-                        name {{
-                          id
-                          nameText {{
-                            text
-                          }}
-                        }}
-                      }}
-                    }}
-                  }}
-                }}
-                """
-            }
-
-            url = "https://api.graphql.imdb.com/"
-            headers = {"Content-Type": "application/json"}
-
-            response = requests.post(url, json=query, headers=headers)
-            data = response.json()
-
-        if response.status_code != 200:
-            return imdb_info
-
-        title_data = self.safe_get(data, ["data", "title"], {})
-        if not data or "data" not in data or "title" not in data["data"]:
-            return imdb_info
-
-        imdb_info['imdbID'] = imdbID
-        imdb_info['title'] = self.safe_get(title_data, ['titleText', 'text'], meta['title'])
-        imdb_info['year'] = self.safe_get(title_data, ['releaseYear', 'year'], meta['year'])
-        original_title = self.safe_get(title_data, ['originalTitleText', 'text'], '')
-        imdb_info['aka'] = original_title if original_title and original_title != imdb_info['title'] else imdb_info['title']
-        imdb_info['type'] = self.safe_get(title_data, ['titleType', 'id'], None)
-        runtime_seconds = self.safe_get(title_data, ['runtime', 'seconds'], 0)
-        imdb_info['runtime'] = str(runtime_seconds // 60 if runtime_seconds else 60)
-        imdb_info['cover'] = self.safe_get(title_data, ['primaryImage', 'url'], meta.get('poster', ''))
-        imdb_info['plot'] = self.safe_get(title_data, ['plot', 'plotText', 'plainText'], 'No plot available')
-        genres = self.safe_get(title_data, ['titleGenres', 'genres'], [])
-        genre_list = [self.safe_get(g, ['genre', 'text'], '') for g in genres]
-        imdb_info['genres'] = ', '.join(filter(None, genre_list))
-        imdb_info['rating'] = self.safe_get(title_data, ['ratingsSummary', 'aggregateRating'], 'N/A')
-        imdb_info['directors'] = []
-        principal_credits = self.safe_get(title_data, ['principalCredits'], [])
-        if isinstance(principal_credits, list):
-            for pc in principal_credits:
-                category_text = self.safe_get(pc, ['category', 'text'], '')
-                if 'Direct' in category_text:
-                    credits = self.safe_get(pc, ['credits'], [])
-                    for c in credits:
-                        name_id = self.safe_get(c, ['name', 'id'], '')
-                        if name_id.startswith('nm'):
-                            imdb_info['directors'].append(name_id)
-                    break
-            if meta.get('manual_language'):
-                imdb_info['original_langauge'] = meta.get('manual_language')
-
-        return imdb_info
-
-    async def get_imdb_info(self, imdbID, meta):
-        imdb_info = {}
-        if int(str(imdbID).replace('tt', '')) != 0:
-            ia = Cinemagoer()
-            info = ia.get_movie(imdbID)
-            imdb_info['title'] = info.get('title')
-            imdb_info['year'] = info.get('year')
-            imdb_info['aka'] = info.get('original title', info.get('localized title', imdb_info['title'])).replace(' - IMDb', '')
-            imdb_info['type'] = info.get('kind')
-            imdb_info['imdbID'] = info.get('imdbID')
-            imdb_info['runtime'] = info.get('runtimes', ['0'])[0]
-            imdb_info['cover'] = info.get('full-size cover url', '').replace(".jpg", "._V1_FMjpg_UX750_.jpg")
-            imdb_info['plot'] = info.get('plot', [''])[0]
-            imdb_info['genres'] = ', '.join(info.get('genres', ''))
-            imdb_info['rating'] = info.get('rating', 'N/A')
-            imdb_info['original_language'] = info.get('language codes')
-            if isinstance(imdb_info['original_language'], list):
-                if len(imdb_info['original_language']) > 1:
-                    imdb_info['original_language'] = None
-                elif len(imdb_info['original_language']) == 1:
-                    imdb_info['original_language'] = imdb_info['original_language'][0]
-            if imdb_info['cover'] == '':
-                imdb_info['cover'] = meta.get('poster', '')
-            if len(info.get('directors', [])) >= 1:
-                imdb_info['directors'] = []
-                for director in info.get('directors'):
-                    imdb_info['directors'].append(f"nm{director.getID()}")
-        else:
-            imdb_info = {
-                'title': meta['title'],
-                'year': meta['year'],
-                'aka': '',
-                'type': None,
-                'runtime': meta.get('runtime', '60'),
-                'cover': meta.get('poster'),
-            }
-            if len(meta.get('tmdb_directors', [])) >= 1:
-                imdb_info['directors'] = meta['tmdb_directors']
-        return imdb_info
-
-    async def search_imdb(self, filename, search_year):
-        imdbID = '0'
-        ia = Cinemagoer()
-        search = ia.search_movie(filename)
-        for movie in search:
-            if filename in movie.get('title', ''):
-                if movie.get('year') == search_year:
-                    imdbID = str(movie.movieID).replace('tt', '')
-        return imdbID
-
-    async def imdb_other_meta(self, meta):
-        imdb_info = meta['imdb_info'] = await self.get_imdb_info_api(meta['imdb_id'], meta)
-        meta['title'] = imdb_info['title']
-        meta['year'] = imdb_info['year']
-        meta['aka'] = imdb_info['aka']
-        meta['poster'] = imdb_info['cover']
-        meta['original_language'] = imdb_info['original_language']
-        meta['overview'] = imdb_info['plot']
-        meta['imdb_rating'] = imdb_info['rating']
-
-        difference = SequenceMatcher(None, meta['title'].lower(), meta['aka'][5:].lower()).ratio()
-        if difference >= 0.9 or meta['aka'][5:].strip() == "" or meta['aka'][5:].strip().lower() in meta['title'].lower():
-            meta['aka'] = ""
-        if f"({meta['year']})" in meta['aka']:
-            meta['aka'] = meta['aka'].replace(f"({meta['year']})", "").strip()
-        return meta
-
-    async def search_tvmaze(self, filename, year, imdbID, tvdbID, meta):
-        try:
-            tvdbID = int(tvdbID) if tvdbID is not None else 0
-        except ValueError:
-            print(f"Error: tvdbID is not a valid integer. Received: {tvdbID}")
-            tvdbID = 0
-
-        if meta.get('tvmaze_manual'):
-            tvmazeID = int(meta['tvmaze_manual'])
-            return tvmazeID, imdbID, tvdbID
-        else:
-            tvmazeID = 0
-            results = []
-
-            if imdbID is None:
-                imdbID = '0'
-
-            if meta['manual_date'] is None:
-                if int(tvdbID) != 0:
-                    tvdb_resp = self._make_tvmaze_request("https://api.tvmaze.com/lookup/shows", {"thetvdb": tvdbID}, meta)
-                    if tvdb_resp:
-                        results.append(tvdb_resp)
-                    else:
-                        if int(imdbID) != 0:
-                            imdb_resp = self._make_tvmaze_request("https://api.tvmaze.com/lookup/shows", {"imdb": f"tt{imdbID}"}, meta)
-                            if imdb_resp:
-                                results.append(imdb_resp)
-                            else:
-                                search_resp = self._make_tvmaze_request("https://api.tvmaze.com/search/shows", {"q": filename}, meta)
-                                if search_resp:
-                                    if isinstance(search_resp, list):
-                                        results.extend([each['show'] for each in search_resp if 'show' in each])
-                                    else:
-                                        results.append(search_resp)
-            else:
-                if int(tvdbID) != 0:
-                    tvdb_resp = self._make_tvmaze_request("https://api.tvmaze.com/lookup/shows", {"thetvdb": tvdbID}, meta)
-                    if tvdb_resp:
-                        results.append(tvdb_resp)
-                if int(imdbID) != 0:
-                    imdb_resp = self._make_tvmaze_request("https://api.tvmaze.com/lookup/shows", {"imdb": f"tt{imdbID}"}, meta)
-                    if imdb_resp:
-                        results.append(imdb_resp)
-                search_resp = self._make_tvmaze_request("https://api.tvmaze.com/search/shows", {"q": filename}, meta)
-                if search_resp:
-                    if isinstance(search_resp, list):
-                        results.extend([each['show'] for each in search_resp if 'show' in each])
-                    else:
-                        results.append(search_resp)
-
-            if year not in (None, ''):
-                results = [show for show in results if str(show.get('premiered', '')).startswith(str(year))]
-
-            seen = set()
-            unique_results = []
-            for show in results:
-                if show['id'] not in seen:
-                    seen.add(show['id'])
-                    unique_results.append(show)
-            results = unique_results
-
-            if not results:
-                if meta['debug']:
-                    print("No results found.")
-                return tvmazeID, imdbID, tvdbID
-
-            if meta['manual_date'] is not None:
-                print("Search results:")
-                for idx, show in enumerate(results):
-                    console.print(f"[bold red]{idx + 1}[/bold red]. [green]{show.get('name', 'Unknown')} (TVmaze ID:[/green] [bold red]{show['id']}[/bold red])")
-                    console.print(f"[yellow]   Premiered: {show.get('premiered', 'Unknown')}[/yellow]")
-                    console.print(f"   Externals: {json.dumps(show.get('externals', {}), indent=2)}")
-
-                while True:
-                    try:
-                        choice = int(input(f"Enter the number of the correct show (1-{len(results)}) or 0 to skip: "))
-                        if choice == 0:
-                            print("Skipping selection.")
-                            break
-                        if 1 <= choice <= len(results):
-                            selected_show = results[choice - 1]
-                            tvmazeID = selected_show['id']
-                            print(f"Selected show: {selected_show.get('name')} (TVmaze ID: {tvmazeID})")
-                            break
-                        else:
-                            print(f"Invalid choice. Please choose a number between 1 and {len(results)}, or 0 to skip.")
-                    except ValueError:
-                        print("Invalid input. Please enter a number.")
-            else:
-                selected_show = results[0]
-                tvmazeID = selected_show['id']
-                if meta['debug']:
-                    print(f"Automatically selected show: {selected_show.get('name')} (TVmaze ID: {tvmazeID})")
-
-            if meta['debug']:
-                print(f"Returning results - TVmaze ID: {tvmazeID}, IMDb ID: {imdbID}, TVDB ID: {tvdbID}")
-            return tvmazeID, imdbID, tvdbID
-
-    def _make_tvmaze_request(self, url, params, meta):
-        if meta['debug']:
-            print(f"Requesting TVmaze API: {url} with params: {params}")
-        try:
-            resp = requests.get(url, params=params)
-            if resp.ok:
-                return resp.json()
-            else:
-                if meta['debug']:
-                    print(f"HTTP Request failed with status code: {resp.status_code}, response: {resp.text}")
-                return None
-        except Exception as e:
-            print(f"Error making TVmaze request: {e}")
-            return None
